requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
loguru>=0.7.2
websockets>=12.0
httpx>=0.25.0
